model.dataset: Add input dimension guard
The number of dimensions for input data should be 2 (channels,
datapoints). Ususally I think there will be only one channel in this
type of data but to generalize the code data should be stored as an
array with shape (1, N_samples). Since this easy to miss, this addition
checks if the shape has one dimension and adds a dimension if required.