train: Add num_anchors_per_scale to yolov3 instantiation
Add the number of anchors per scale to the yolo network. This is
necessary when using multiple anchors for one scale. The number of
anchors for one scale is determined automatically from the config. The
config is assumed to have an anchor parameter which is a list of lists
with the second dimension equal.