Add augmentnation module
Add a module wich can apply random augmentations. The augmentations
module is loosely inspired by the Albumentations package, however its
implementation is far simpler.

An augmentation is a change to the input vector which changes the vector
in some way (some transformation like translation, flipping etc.) which
also changes the bounding domain/box of a label accordingly.

Augmentations can be randomly generated using the Distribution classes
provided in the module. Currently normal and gaussian distributions are
implemented. The build on top of numpy.random distributions and are
therefore seedable by seeding the numpy.random module.

The input domain and corresponding labels are assumed to be on a
normalized domain of [0, 1]. Labels are assumed to have the format
(mean, width, class). Currently, no other formats are supported.

This commit adds the following augmentations:
1. Translate: shift the data on domain. Values will be extrapolated
2. Flip: flip the data along the y=0.5 line
3. PolynomialBaseline: Add a polynomial baseline curve to the data

Closes issue #10