model.yolov3: rework route connection system
Previously route connections were hardcoded to occur when the residual
block had 8 repeats. This is an issue if we want to reduce the model
size. For this reason, the config has been reworked such that the
strings in the residual block config list objects now has a meaning. The
string "R" now stands for residual block. The string "S" stands for skip
connection (i.e. route connection).

When the layers are being built, the method will now check which strings
are present in the 0th index of the list. Multiple strings can be
present at once. The check works by checking if the identifier token
("R" and/or "S") are present in the proveded string.

Supplying the block ["R", 2] will add 2 residual blocks. ["RS", 3] will
add 3 risidual blocks followed by a skip connection. If "R" and "S" are
supplied in whatever order, the residual block is always built first.
Since the location index of the skip connection is equal to that of the
module list index, a skip connection may only be added after a residual
block (otherwise the indexation will be messed up!). This should be
fixed at some point in the future.

Also removes some deprecated comments and variables left over from
debugging and adds a docstring.