Better build config (#21) (#26)
* model.yolov3: rename config to default_config

As more configs are to be added, the config object will become the
"default_config" which mimicks most closely the standard (2D) yolov3
network.

* model.yolov3: Fix config issues in Yolo1DV3 network

Fix an issue where the self.config was set after creating the layers
(which requires self.config), and add self.config in the
_create_conv_layers method which previously accessed the (now renamed)
`config` global variable directly.

* model.yolov3: remove deprecated comments

The removed comments were added during debugging and no longer apply.

* model.yolov3: Add a YoloConfigException

This exception handles issues pertaining to incorrect config formatting
resulting in parsing issues on runtime.

* model.yolov3: rework route connection system

Previously route connections were hardcoded to occur when the residual
block had 8 repeats. This is an issue if we want to reduce the model
size. For this reason, the config has been reworked such that the
strings in the residual block config list objects now has a meaning. The
string "R" now stands for residual block. The string "S" stands for skip
connection (i.e. route connection).

When the layers are being built, the method will now check which strings
are present in the 0th index of the list. Multiple strings can be
present at once. The check works by checking if the identifier token
("R" and/or "S") are present in the proveded string.

Supplying the block ["R", 2] will add 2 residual blocks. ["RS", 3] will
add 3 risidual blocks followed by a skip connection. If "R" and "S" are
supplied in whatever order, the residual block is always built first.
Since the location index of the skip connection is equal to that of the
module list index, a skip connection may only be added after a residual
block (otherwise the indexation will be messed up!). This should be
fixed at some point in the future.

Also removes some deprecated comments and variables left over from
debugging and adds a docstring.

* model.yolov3: change config to use "RS" format

default config now uses "RS" format to add skip connections.

* model.yolov3: Add tiny model architecture and tests

Add tiny architecture and add a build test. The tiny architecture has
fewer repeats of the residual blocks but is otherwise the same. This
reduces the number of model parameters.

* model.config: Add BUILD_CONFIG variable

Add the BUILD_CONFIG variable which allows the user to easily select
which config to use from the yolov3 configs.

* train: Add config.BUILD_CONFIG to train script

The config.BUILD_CONFIG variable is now used to set the models build
config in the training script.

---------

Co-authored-by: Luc Kusters <luc.kusters@anatech.nl>