model.dataset: fix indentation error
Fix incorrectly indented return of Dataset.__getitem__ which was
previously inside a for loop loping over anchor_indices, thereby not
allowing the loop over all anchor indices. This caused an issue where
targets with multiple labels were only returned with one label due to
the early exit of the loop, since all the targets are assigned to one of
the anchor boxes but only one anchor box could be assigned.

closes #30