model.dataset: Fix typo and add copy
Fixes a typo (name only, not leading to errors).

Also makes a copy of series when an augmentation is applied. This is
needed when the flip augmentation is applied as it introduces negative
stride, which is not supported in PyTorch. It might actually be better
in the future to add copy only when applying the flip transformation
however. This is a catch-all solution but might be a bit less memory
efficient.