model.dataset: make anchor_on_scale toggleable
Tests showed that removing the anchor_on_scale condition improves
confidence, recall and precision of the network. It seems to me that
this might be a superfluous check. For this reason I'm changing this to
a togglable variable, off by default.

The only downside of doing this is that now there seem to be many more
target labels, which is messing with mAP and visualization of target
labels.